+++ src/utils/listPaths.js
@@ -4,8 +4,10 @@ import fs from 'fs';
 import path from 'path';
 import globby from 'globby';
 
-const listPaths = async (srcPatterns: string): Promise<Array<string>> => {
-  const paths = await globby(srcPatterns);
+const listPaths = async (src: string, ignoreSrc?: ?string): Promise<Array<string>> => {
+  const patterns = [src];
+  if (ignoreSrc) patterns.push(`!${ignoreSrc}`);
+  const paths = await globby(patterns);
   return paths.filter((filePath) => {
     try {
       return fs.statSync(path.resolve(process.cwd(), filePath)).isDirectory()