+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Add `--ignore-src <glob>` option to all commands to exclude sub-packages (#38)
 * Add warning to `oao oudated` for internal deps that do not meet the specified version range (#34)
 
 ## 0.8.0 (Mar. 8, 2017)